Fix broken scrolling at bottom of window
The set-window-start bit seemed to make it such that, when doing
newline-and-indent on the last line of a window, the newline would get
inserted, the following line would get indented properly (I think),
but my window wouldn't scroll down to show the new line.  Instead my
cursor would jump to the center of the window.

This hasn't broken anything for me yet, which I'm a little surprised
about.  YMMV, I reserve the right to revert/improve this later if I
find out that this code I removed was actually necessary for
something.