Rebind indentation variables instead of defalias
defalias bitches if your variable is local (or can become local?).  It
may be overkill, anyway.  I'm going to try just rebinding the variables
to all be the same as tab-width right before we call the real indent
function and see how this serves me.  This change may have unwanted
consequences, though.  More testing is needed.